Neo can now identify its location
Added an entityData class that can be attached to every player in order
for Neo to track their status and location. Neo answers the question 'where are you' by checking its own status in the DB. The code is kind of dirty since I'm
trying to get the unit tests to pass as quickly as possible. I will clean
up later as I add more exception handling.