Neo can now go to different rooms!
Added tv and couch objects to main room

Added navmesh agent for neo, I had some issues with the onTriggerEnter
event firing hundreds of times per second when neo entered the boundary
box of a room, so unfortunately I had to remove the boxes and replace them
with plane objects in front of the entrance to each room. It's not ideal
but it will work for this prototype.

added "go" command for wernickeArea which finds the coordinates of the
given location and tells neo to go there. I had issues with parsing
locations with multiple words (ie "main room") so I threw in a temp
solution which will have to be improved upon.

For now though, Neo can go into any of the rooms by command.