Neo memorizes object coordinates
Neo will now memorize the x and z coordinates for objects
when storing their meta data. This will make it easier for us to
get it to track objects and move to their respective locations.

I'd like to personally thank my past self for creating such useful
methods for extracting the meta-data from the DB. It made this
commit so much easier to complete and test this milestone.