Schema change to DB
I decided it would be easier to store the x and z coordinates as columns
in the objects table rather than the attributes table because each object
should have its on unique coordinates. This makes it easier for extraction
and organizing the info in general.

Hopefully I won't have to make many more changes to the schema for VER 1.0
I've added sql scripts for generating the DB schema as it is right now in
case I make more changes that jack stuff up.