Improve db utils to make easier to use (#1)
* Made DBUtils class a little easier to use

Before this commit, some of the DBUtil methods required you to pass in a
IdbCommand object which made it a little difficult to work with. This
commit has all the DBUtil methods handle creating the dbCommand object so
the caller just has to pass in the pieces of data.

* fix unit tests to use correct db files

with all the refactoring done to the dbutils class, this commit fixes the
unit tests so their calls only go to the test db files.