Added Statement class to improve API usability
In order to make it easier for developers to create new natural language
commands for the agent to process, this commit adds the wordType enum and
the Statement class which can be used to quickly create new statements
that the agent can recognize and respond to. Read the Statement class
comment block for more info on how it works.